Runs GSEA on a pre-ranked list of differential expression statistcis with fgsea

Techinically this doesn't require a full design to run, but it proves expedient for a quick turnaround.

do.fgsea(
  gsd,
  x,
  design,
  contrast = ncol(design),
  minSize = 15,
  maxSize = 15,
  nperm = 10000,
  gseaParam = 1,
  score.by = c("t", "logFC", "pval"),
  logFC = NULL,
  gs.idxs = as.list(gsd, active.only = TRUE, value = "x.idx"),
  ...
)

Arguments

gsd
The GeneSetDb() for analysis

x
The expression object. This can be 1 column matrix if you are not running any analysis, and this function essentially is just a "pass through"

design
The design matrix for the experiment

contrast
The contrast you want to test and provide stats for. By default this tests the last column of the design matrix. If you want to test a custom contrast, this can be a contrast vector, which means that it should be as long as ncol(design) and it most-often sum to one. In the future, the user will be able to specify a range of coefficients over design to perform an ANOVA analysis, cf. Issue #11 (https://github.com/lianos/multiGSEA/issues/11).

...
arguments to pass down into calculateIndividualLogFC()

Value

A data.table of fgsea results.

Details

Note that fgsea isn't added to import or suggests because rescomp's compiler can't handle this yet.

fgsea expects the pathways to be describes as a list of character vectors where the vectors are gene ids, and the names of the pre-ranked vector correspond to those IDs

This function is not meant to be called directly. It should only be called internally within multiGSEA().