Fetches information for a gene set

Gene sets inside a GeneSetDb() are indexed by their collection,name compound key. There is no special class to represent an individual gene set. Instead, gene sets are returned as a data.frame, the rows of which enumerate the features that belong to them.

When x is a MultiGSEAResult(), this function will append the differential expression statistics for the individual features generated across the contrast that defined x.

geneSet(x, i, j, ...)

# S4 method for GeneSetDb
geneSet(
  x,
  i,
  j,
  active.only = is.conformed(x),
  with.feature.map = FALSE,
  ...,
  collection = NULL,
  name = NULL,
  as.dt = FALSE
)

# S4 method for MultiGSEAResult
geneSet(
  x,
  i,
  j,
  active.only = TRUE,
  with.feature.map = FALSE,
  ...,
  collection = NULL,
  name = NULL,
  as.dt = FALSE
)

Arguments

x
Object to retrieve the gene set from, either a GeneSetDb or a MultiGSEAResult.

i
The collection,name compound key identifier of the gene set

j
The collection,name compound key identifier of the gene set

...
passed down to inner functinos

active.only
only look for gene sets that are "active"? Defaults to TRUE if x is conformed to a target expression object, else FALSE. conform() for further details.

with.feature.map
If TRUE, then details of the feature mapping from the original feature_id space to the target feature space are included (default: FALSE).

as.dt
If FALSE (default), the data.frame like thing that this funciton returns will be set to a data.frame. Set this to TRUE to keep this object as a data.table

Value

a data.(frame|table) of gene set information. If x is a MultiGSEAResult object, then differential expression statistics are added as columns to this result.